Fix return arg in tribits_create_name_from_current_source_directory() (#299)
Wen I did the refactoring of tribits_add_executable() to break out some
subfunctions, it exposed a defect from the following commit 13 years ago:

  654a4ef8843 "     SUMMARY:   Cmakify Intrepid"
  Author: Brent M. Perschbacher <bmpersc@sandia.gov>
  Date:   Thu Apr 2 15:54:30 2009 +0000 (13 years ago)

The problem was an incorrect setting of an output function argument:

  FUNCTION( PACKAGE_CREATE_NAME_FROM_CURRENT_SOURCE_DIRECTORY DIRECTORY_NAME )
    SET(DIRECTORY_NAME "")
    ...
    SET(DIRECTORY_NAME ${DIRECTORY_NAME} PARENT_SCOPE)
  ENDFUNCTION()

When I renamed the variable `DIRECTORY_NAME` in the parent scope, it broke
this function (because DIRECTORY_NAME was no longer being used).  (If you are
going to just set a var in the parent scope unconditionally, why bother
passing in some name in the function?  Just another example of where
inexperienced developers have problems with the strange CMake language.)

NOTE: That function got renamed and moved around but I was able to track it
along the CMake history.