Support list for TAAT() PASS_REGULAR_EXPRESSION and FAIL_REGULAR_EXPRESSION (#464)
Turns out that the tribits_add_advanced_test() test-block arguments
PASS_REGULAR_EXPRESSION and FAIL_REGULAR_EXPRESSION did not support taking in
a list of regexes like the built-in CMake CTest properties of the same names.
This commit updates TAAT() to allow a list of arguments and they behave the
exact same way as the built-in properties (i.e. matching any).

I also updated the documentation to show how to pass in a list of regexes
correctly for PASS_REGULAR_EXPRESSION, FAIL_REGULAR_EXPRESSION,
FINAL_PASS_REGULAR_EXPRESSION and FINAL_FAIL_REGULAR_EXPRESSION.

I also fixed the documentation for tribits_add_test() as well for
PASS_REGULAR_EXPRESSION and FAIL_REGULAR_EXPRESSION.

Basically, a well constructed CMake function should almost never take in a
list of arguments expecting an explicit semi-colon.  This was part of the
learning in PR #464 with how to treat semi-colons in CMake.

This commit was driven by a usage of FAIL_REGULAR_EXPRESSION in Trilinos while
testing updated TriBITS as part of #299 when I realized the documentation was
wrong for how to deal with these lists of regexes.