Snapshot FindTPLNetcdf.cmake from Trilinos 8c1028df724 with small mods (#299)
This was snapshotted from the Trilinos  commit:

  commit 8c1028df724eb0096cd688b381285112bc6f914a
  Author: Roscoe A. Bartlett <rabartl@sandia.gov>
  Date:   Wed May 18 15:08:17 2022 -0600

    FindTPLNetcdf.cmake: Lower-case function names (TriBITSPub/TriBITS#274)

soon to be merged to Trilinos 'develop' from Trilinos PR #10533.

However, this is not an snapshot as I made the following changing before
creating this commit:

* Kept the spelling fix for 'separated' made in TriBITS 'master' branch from
  commit c716a74e8419d356fd16552eaf74ece85397eccd; Author: Greg Sjaardema
  <gsjaardema@gmail.com>; Date: Mon Aug 2 16:02:20 2021 -0600; "Spelling
  fixes"

* I removed a few trailing spaces at the end of some lines

Why? There really is no value in using a different version of
FindTPLNetcdf.cmake in TriBITS that what is used in Trilinos.  When testing
Trilinos against updated versions of TriBITS, you really want the logic to
match.  In the case in testing with Trilinos (as part of #299 and testing with
trilinos/Trilinos#10533), the old version of FindTPLNetcdf.cmake before this
change was not properly setting TPL_Netcdf_PARALLEL which was not allowing the
enable of the test SEACASIoss_exodus_fpp_serialize.  With this sync, we get
the exact same tests.