WIP: Set <Project>_INCLUDE_DIRS in <Project>Config.cmake for the install tree (#299)
Turns out that Albany explicitly looks for files under
${Trilinos_INCLUDE_DIRS}.  Also, Albany was looking for files like
"${Trilinos_INCLUDE_DIRS}/Kokkos_config.h" so it basically requires
Trilinos_INCLUDE_DIRS to be just a single directory!

Therefore, to avoid having to change Albany, I have set <Project>_INCLUDE_DIRS
in the installed <Project>Config.cmake file.  This was easy to do for this
case.

But note that <Project>_INCLUDE_DIRS in <Project>Config.cmake in the build dir
still will be set to empty because there is no easy way to set it in that
case.

This is a stop-gap measure to allow APPs like Albany to keep working during
this initial transition to modern CMake.  But we really should remove this var
from the <Project>Config.cmake file entirely at some point.

NOTE: I have not fixed all of the TriBITS tests for this change yet.  That is
why this is a 'WIP' commit.  I will fix the tests in a later commit.