Produce CDash URLs to all builds and nonpassing tests for same repo version (#483)
This makes it easy to click on links to these other CDash URLs to see what is
happening with all of the builds for the same repo version.

The main target is for the GitHub Actions drivers for TriBITS testing.  This
makes it easy to get at those results.

As part of this, I also renamed the function
tribits_get_build_url_and_write_to_file() to tribits_get_cdash_build_url() and
removed the ability to write to a file.  (We just don't need that anymore and
it is trivial to write a string to a file.)