Have 'dashboard' target set CTEST_UPDATE_VERSION_ONLY=TRUE
This will post the base repo version info to CDash which provides more info
right on CDash on the main index.php page.

However, it will set CTEST_DO_UPDATES=OFF unless the base repo is a Git repo
and we can extract the SHA1 from the repo with tribits_git_repo_sha1() (which
is a very robust function).