Only link to <Package>_TEST_ENABLED_DEPENDENCIES for TESTONLY libs (#299)
It would be bad to link non-testonly libs to these libraries because that
would violate the TriBITS dependency handling system.  It would allow a
package to depend on a package listed as a test dependency and not listed in
the lib dependency.  And various algorithms like enabling downstream package
and tests would be incorrect.

Note that I actually had a test that showed this problem.  The executable
c/c_util was showing linking against libmixedlang.a but none of the
dependencies of that (sub)package or its upstream packages have a lib
dependency on the MixedLang package.  This library was coming in through the
TEST dependency on MixedLang in the package WithsubpackagesB.  I was sloppy in
just accepting the link line in this case as I did not carefully manaully
inspect the package lib dependencies.