Remove quotes from --pretty argument value (#483)
In this case, there are no spaces in the --pretty=format:<value> argument so
we don't need the quotes and therefore don't need to suffer the complexity of
removing them after the fact.

But note the case I copied and pasted this from in
tribits_generate_single_repo_version_string() still requires them because that
format spec has spaces.