Call cmake_policy() at file-scope and not in functions (#485)
It was pointed out by @KyleFromKitware in the review of PR #485 that calling
cmake_policy() in a function will change a policy in the calling scope.

NOTE: The function() command does not create a new policy scope but does
create a new variable scope.  Just the opposite, include() creates a new
policy scope but not a new variable scope.