Remove assert for TPL_<tplName>_LIBRARIES (#299)
With the new TriBITS TPl implementation, the is no need for a
FindTPL<tplName>.cmake file to set the var TPL_<tplName>_LIBRARIES so there is
no need to assert it.

This came up when refactoring the FindTPLCUBLAS.cmake, FindTPLCUSPARSE.cmake
and FindTPLCUSOLVE.cmake files in Trilinos as part of the PR
trilinos/Trilinos#10614.