Allow lib files not beginning with 'lib' on APPLE, improve err msg (#299)
* Added unit tests to make sure that MacOSX framework files
  (e.g. /some/base/path/Accelerate.framework) is allowed on MacOSX
  (i.e. APPLE===TRUE)

* Improved error message for invalid lib name

* Added unit tests for APPLE cases that we know we need to support

This was driven by testing on MacOSX as part of PR trilinos/Trilinos#10614
where it was reported (offline, not in the issue) that using
find_package(BLAS) will find the file /some/base/dir/Accelerate.framework on a
Mac.  Hopefully CMake will not have a problem with an imported target with the
file path Accelerate.framework.  (I am hoping that this just gets CMake to put
that file in the link line in the correct order.)