WIP: Add failing test for find_package( ... OPTIONAL_COMPONENTS ... ) (#299)
The tests fail for this commit because it is showing a defect in TriBITS!
This was reported by the Xyce team with testing against
trilinos/Trilinos#10614.

TriBITS was missing a test for find_package(<Project> ... OPTIONAL_COMPONENTS
...) for the <Project>Config.cmake file and the handling of these was broken.
The code is successfully only including the found optional components in the
var <PROJECT>_LIBRARIES but it was seting <PROJECT>_SELECTED_PACKAGES_LIST and
<Project>::all_libs for all of the required and optional components,
regardless if the optional components were found or not.

What I did:

* Upgraded the existing tests based on the DummyPackageClientCMakeLists.txt
  project to call find_package(<Package> REQUIRED COMPONENTS
  ... OPTIONAL_COMPONENTS ...) and accept the list of optional components.
  (This required updating several tests that were using this.)

* Added a new failing test case TEST_7 that showed the incorrect handling of
  OPTIONAL_COMPONENTS in that it was not excluding optional components that
  were not found by not excluding them from the
  <Project>_SELECTED_PACKAGES_LIST variable or from the
  <Project>::all_selected_libs targets.

NOTE: The new test case TEST_8 shows that you can pass in subpackages as
components and it will handle that correctly as well.

Other little stuff I did:

* Updated some of the other vars in TribitsProjectConfigTemplate.cmake.in

* Fixed some comment typos I happened to notice

* Added some more comments about what things are being tested in various tests

* Removed some redundancy in testing the <Project>Config.cmake and
  <Package>Config.cmake files

* Added seperator comments between some long test code