WIP: Update TribitsExampleProject2 tests to expose defect in installed <tplName>Config.cmake files (#500)
The updated some tests to copy in the source dir, build and install in
different subdirs, and delete the soruce and build dir.  Then the downstream
TribitsExampleApp2 tests configure and build against the install dirs where no
src or build dir exists.

And with this, we see the defect in the installed <tplName>Config.cmake
showing them pointing into the build dir for find_dependency() calls called
out in (#500).

NOTE: The 'find_package' tests did not show any errors because they ignore
upstream TPL dependencies.  They just set <externalPkg>_DIR to what was found
by the call to find_package(<externalPkg>) in the FindTPL<tplName>.cmake
module.  The generated <tplName>Config.cmake file sets <externalPkg>_DIR to
that same dir and then calls find_dependency(<externalPkg>).  The assumption
is that <externalPkg> is found outside of the build dir and will be valid for
installs as well.

NOTE: I also moved ALWAYS_FAIL_ON_NONZERO_RETURN to the bottom of TEST_<IDX>
blocks to make them more noticable (and that is typically where I put them).