Fix <tplName>Config.cmake files to not point into build dir (#500)
Now, this sets <upstreamTplName>_DIR in Tribits wrapper files to set
<upstreamTplName>_DIR to "${CMAKE_CURRENT_LIST_DIR}/../<upstreamTplName>}.
This will work for the the build tree and the install tree.  It also makes the
install relocatable, w.r.t. these files.

Note that we did not change <externalPkg>_DIR in the <tplName>Config.cmake
file generated using the function
tribits_extpkg_create_package_config_file_with_imported_targets().  That is
because those are for external packages that should not be in the build tree
but should instead be in the install tree.  To make those
<tplName>Config.cmake files relocatable would be more work (where we would
need to compute relative paths and use those instead when asked).