Correctly handle abs dir for findmod for TPL dependencies file (#10774)
This fixes a defect added as part of work on TriBITSPub/TriBITS#299.  This
makes the failing tests TriBITS_TribitsExampleProject2_find_tpl_parts from the
previous tests pass.

I also updated the documentation to mention that <tplName>_FINDMOD and
<tplName>_DEPENDENCIES_FILE can also be abs dirs.

And I checked carefully, if these are relative paths, then the variables are
relative to <projectDir>, not <repoDir>.  But the <findmod> field in the call
to tribits_repository_define_tpls() is relative to <repoDir>.  There is a
translation of the <findmod> field to the variable <tplName>_FINDMOD.