Replace include_guard(GLOBAL) with if (NOT TARGET ${depPkg}::all_libs) (#505)
This fixes the test added in the last commit for calling find_package() for
related packages from different subdirs that was failing.

I belive the problem is that the targets created by the CMake export() command
don't have GLOBAL on them and only have directory scope.  Therefore, we can't
use include_guard(GLOBAL) because that will stop the creation of the imported
targets in a different subdir where they are not defined.