Change from 'http' to 'https' for CDash submissions
After a recent change to the testing-dev.sandia.gov/cdash/ site, you now need
to submit with 'https' instead of 'http'.  This works for the site
testing.sandia.gov/cdash/ as well.

I also tested by configuring and building with:

  -DTriBITS_CTEST_DRIVER_SUBMIT_TO=TESTING_SANDIA_CDASH

which submitted to:

   https://testing.sandia.gov/cdash/index.php?project=TribitsExProj&date=2022-08-05&showfilters=1

and also:

   https://testing-dev.sandia.gov/cdash/index.php?project=TribitsExProj&date=2022-08-05&showfilters=1

due to:

-- ENV_TRIBITS_2ND_CTEST_DROP_SITE='testing-dev.sandia.gov'
-- TRIBITS_2ND_CTEST_DROP_SITE='testing-dev.sandia.gov'

NOTE: I need to namespace these variables by the project name so as to avoid
this type of cross-talk in the future.