Fix test dependencies for subpackages and subpackage tests/examples enabless (#63, #268, #299, #510)
This fixes a defect added as part of transitioning to modern CMake targets
(#63, #299).

As part of this, it also changes the behavior of TriBITS to only enable
tests/examples for subpackages that are enabled on the forward sweep and not
those that are enabled as part of sweeping backward to enabled upstream
dependencies needed by enabled packages (see #268).  This is the correct thing
to do and will also result in fewer tests being being built and run as part of
testing downstream packages using:

  -D<Project>_ENABLE_<Pkg>=ON
  -D<Project>_ENABLE_ALL_FORWARD_DEP_PACKAGES=ON
  -D<Project>_ENABLE_TESTS=ON

These changes causes failing tests in the previous commit to pass.

I also updated some documentation and removed some commented-out debugging
code.