Add version check and calls to cmake_minimum_required() (#522)
This addresses an issue where Kokkos updates the KokkosConfig_install.cmake
file to use 'if (... IN_LIST ...)'.

NOTE: There is no test in TriBITS that checks for this change because the
TriBITS-generated <Package>Config.cmake and <Project>Config.cmake files don't
use 'if (... IN_LIST ...)'.  I had to manually test this against Trilinos to
verify this makes the error go away.  All that the TriBITS testing does is
ensure that this change does not break these files for the TriBITS use cases.