Change cmake_minimum_required() to upper bound TRIBITS_CMAKE_MINIMUM_REQUIRED (#522)
This was changed from the open-ended ${CMAKE_VERSION}.  This way, we allow
CMake to use a more current set of policies while not writting future futures
of CMake a blank check.

This change was made based on feedback from @KyleFromKitware from the
post-merge review of #523.

NOTE: I manaully tested this with a configure of Trilinos and it produded:

  cmake_minimum_required(VERSION 3.3...3.17.0)