Document undocumented CMake -C *.cmake FORCE set() behavior
This adds a new bullet to document surprising behavior of set( ... FORCE)
statements in *.cmake files read in using -C vs. the same statements that
appear in the projects CMakeLists.txt files (or pulled in through the TriBITS
option <Project>_CONFIGURE_OPTIONS_FILE).

I also added a small sub-table-of-contents for the three different configure
options that exist.