Don't set <externalPkg>_DIR when using a find module (#299)
When find_package(<externalPkg>) uses a find module Find<exteranlPkg>.cmake
instead of a package config file <externalPkg>Config.cmake, the variable.
<externalPkg>_DIR is not set.  Therefore, it is best to not set
<externalPkg>_DIR in the TriBITS-geneated <tplName>Config.cmake file since
that is just confusing.

NOTE: The case where TriBITS a FindTPL<tplName>.cmake module calls
find_package(<externalPkg>) is the only a situation where this is the case.
When one TriBITS TPL is calling find_dependency() on an upstream TriBITS TPL,
the <upstreamDepTpl>_DIR variable is always set.