Document the need for <Package>_FULL_ENABLED_DEP_PACKAGES (#299, #63)
Unfortunately, I could not remove the computation of the full set of direct
and indirect dependencies for each pacakge <Package>_FULL_ENABLED_DEP_PACKAGES
because the function tribits_find_most_recent_file_timestamp() still needs
this.  The only customer that still needs that function is CASL VERA (to
compute when it is needed to rebuild MOOSE).

Therefore, for now, I can't remove this code (which is more complex and can be
an expensive computation if there are a **lot** of packages).