Reorder and group 2nd-level macros/functions (#63)
This is using a form of "Clean Code" ordering where the top-level
macros/functions are at the top and the macros/functions that they call are
below them.

NOTE: I purposefully did not update the foreach() loops involving the list
vars `<Package>_[FORWARD_][LIB|TEST]_[OPTIONAL|REQUIRED]_DEP_[TPLS|PACKAGES`
because all of those loops are going to be replaced with the new vars
`<Package>_[LIB|TEST]_DEFINED_DEPENDENCIES` and
`<Package>_[LIB|TEST]_DEP_REQUIRED_<depPkg>` in the next topic-branch.