WIP: Refactor tribits_sweep_backward_enable_upstream_packages() to process TPLs (#63)
* Added calls to tribits_get_package_enable_status() to make work for internal
  and external packages.

* The message printed when enabling a TPL with the combined logic is different
  so the tests had to change.

* Fixed logic in tribits_set_up_enabled_only_dependencies() for package's with
  index '0' that was handled incorrectly.

* Had to set ${PROJECT_NAME}_DEFINED_PACKAGES in unit test driver because this
  is now used in tribits_adjust_package_enables().

* Change name of TPL_<tplName>_ENABLING_PKG to <tplName>_ENABLING_PKG and set
  for all packages, not just for packages triggering the enable of required
  TPLs.  (Was needed for info on why a TPL got enabled so a user understands.)

* Removed some commented-out debug code.

WARNING: This does **not** handle the implicit enable of upstream TPLs
correctly.  With this current implementation, if a package depends on LAPACK,
for example, TriBITS will automatically try to enable BLAS.  That is not
really the behavior that we want (yet) and breaks backward compatibility.  I
will fix that in a later commit.  (It just so happens that the entire TriBITS
test suite passes so that means we are missing a test for this use case.)