WIP: Update setting of CPACK_COMPONENT_<upstreamComp>_DEPENDS to use new deps vars (#63)
This is not really updated correctly because it include exteranl dependencies
as well but this does not break any TriBITS packages and I am not sure who is
using component-based installs with TriBITS projects.

This can be fixed correctly once I determine that it is needed and we can
develop a test for this.