Remove unused module CMakeOverrides.cmake (#429)
The file CMakeOverrides.cmake contained tribits_include_directories() and the
(terrible) override of include_directories().

It is unclear from TriBITS git history when the mdoule
TribitsIncludeDirectories.cmake and why CMakeOverrides.cmake was not deleted
but there are no current includes of the module "CMakeOverrides" in TriBITS
anywhere.  It appears that the include of the module CMakeOverrides.cmake was
removed 9 years ago and replaced with the include of
TribitsIncludeDirectories.cmake way back in the commit:

  cd8e3e82 "Moving TriBITS-specific funtion into tribits/package_arch/"
  Author: Roscoe A. Bartlett <bartlettra@ornl.gov>
  Date:   Mon Mar 10 16:21:51 2014 -0400 (9 years ago)

So nice to get rid of dead code!