Suppress deprecation warnings for unit tests of deprecated functionality (#429)
The only remaining deprecation warnings are for tests that run an grep the
STDOUT for the deprecation warning text and test
TriBITS_CTestDriver_TribitsExMetaProj_version_date that pulls in older
versions of the repos (that still use include_directories()).

In a later commit, I will update the snapshots of the TriBITS example repos to
update the test TriBITS_CTestDriver_TribitsExMetaProj_version_date to make
these deprecation warnings go away.