WIP: Change printing of package set lists based in INTERNAL/EXTERNAL (#63)
Now printing of the final sets of internal and external packages is based on
<Package>_PACKAGE_BUILD_STATUS and so initially internal packages that are
treated as external packages are not listed under:

  "Final set of [enabled|non-enabled] [toplevel] packages:"

but instead are listed under:

  "Final set of [enabled|non-enabled] external packages/TPLs:"

Hopefully this will make it more clear to the user how these packages are
being treated.

The new tests were updated to adjust to this different output.  (NOTE: We
could remove the direct printing and checking of
<Package>_PACKAGE_BUILD_STATUS but I will leave that in for now.)

This also helped to remove some code by doing:

* The functions tribits_print_internal_[toplevel_]package_list_enable_status()
  were renamed to tribits_print_[toplevel_]package_list_enable_status() and
  the argument internalOrExternal was added.

* The function tribits_print_tpl_list_enable_status() was deleted.