WIP: Add basic logic for setting <Package>_PACKAGE_BUILD_STATUS=EXTERNAL (#63)
This sets an INTERNAL package to EXTERNAL if TPL_ENABLE_<Package>=ON or if a
downstream dependent package is EXTERNAL.

NOTE: This is missing the changes needed to affect the processing of internal
vs. external packages but this should complete the dependency-based logic on
what packages need to be treated as EXTERNAL based on the setting of
TPL_ENABLE_<Package>=ON.

This also changes the logic so that the backward sweep to enable upstream
external packages only does so explicitly if the downstream dependent package
has <Package>_SOURCE_DIR!="" so it if it was initially defined as an internal
package then it will not change the enable/disable behavior.  This is
important because we don't (yet) want an external TPL enable to trigger the
enable of an upstream TPL.  We only want that logic to apply to initially
internal packages.  (The dependency logic can be pretty confusing if we are
not careful.)

I also removed the documentation that setting:

  -D <PackageTreatedAsExternal>_ROOT=<path>

will make an initially internal package be treated as an external package/TPL.
I think that is too subtle.  The user can still set to change where
find_package(<PackageTreatedAsExternal>) will find the
<PackageTreatedAsExternal>Config.cmake file, but that will be the only impact
of setting <PackageTreatedAsExternal>_ROOT.  (We may change our minds on this
latter and the logic for making that decision is pretty encapsulated.)

I had to make some tweaks to get this logic to work correctly:

* tribits_get_package_enable_status(): Changed logic to first just look for
  non-empty vars ${PROJECT_NAME}_ENABLE_${packageName} and
  TPL_ENABLE_${packageName} to determine if the package should be treated as
  internal or external.  If both of those are null, then use
  ${packageName}_PACKAGE_BUILD_STATUS.  This is needed for the logic where
  ${packageName}_PACKAGE_BUILD_STATUS is not changed from INTERNAL to EXTERNAL
  until later in the process.