Make WrapExernal work with external TribitsExProj packages (#63)
So for there is just one test for this.

To make this work, I had to grab the include directories off of the upstream
targets, which are returned as generator expressions.  And with that, I had to
switch the generation of the makefile for external_func to build time with a
custom command so it would evaluate the generator expressions.  (There is no
way to run a general command at generation time.)