Add basic special handling for fully TriBITS-compliant external packages (#63)
Now we don't need to find upstream TPLs already found in upstream TriBITS
package configure and build.  In this case, we don't need to find
HeaderOnlyTpl or SimpleTpl already found in the configure and build of
SimpleCxx.

But this does not yet support stagged sets of package installs.  That will
come next.