Process TPLs at base project level, pass info through <Package>Config.cmake files (#63)
This changes to process TPLs at the base project level instead of in a
function() scope so that local vars set in find_package() calls are available
in the base project scope.  This technically breaks backward compatibility but
this is more similar to how a raw CMake project would call find_package() to
get external packages.  Hopefully this will not break anything.

To accomplish this, this changed the function tribits_process_enabled_tpl() to
a macro.  Also, the macro tribits_process_enabled_tpl() was refactored into
smaller macros and functions to make it easier to understand and maintain.

To test this, this commit includes changes to TribitsExampleProject to pass
info about expected dependencies through the <Package>Config.cmake files to
downstream packages instead of using the hacked ExpectedDepsStrings.cmake
file.  (For many reasons, a TriBITS package should not refer to anything
outside of the base package source dir.)  This fixes all of the tests in
TriBITS.