Initial implementation of pre-building/pre-installing stages of packages (#63)
This commit in an initial completion of for sub-use cases for #63 "Use Case 3:
Configure/build pointing to a subset of already installed TriBITS packages in
same repo".

To make the initial tests for the TribitsExampleProject project to pass, this
commit geneates wrapper <Package>Config.cmake files for TriBITS-compatible
external packages.  This provides the info needed for downstream internal
TriBITS package's <Package>Config.camke files to get the upstream
TriBITS-compatible external package.

This commit makes the earlier test:

  TriBITS_TribitsExampleProject_External_Package_by_Package

fully pass.

However, we need to still add more tests, documentation and error checking but
at least this is a start and all of the tests pass.