gitdist.py: Remove #print(...) comments, resolve codespell issue (#571)
While running codespell 2.2.1, it flagged 'n{nothing}Does' in:

  #print("\n{nothing}Does ...")

which is not a misspelling (and a later version of codespell does not flag
this) but I wanted to remove it since it was not really being used.

And while I was at it, I remove all of the commented out print() statements
(since those were just there for debugging).