Link to correct <Package>Config.cmake files for TRIBITS_PKG files (#63)
This is needed for KokkosConfig.cmake generated by native Kokkos CMake build
system that is not fully TriBITS compliant.

NOTE: This is currently lacking a native TriBITS test but this was validated
with Kokkos+Trilinos.  Some more functionality will be needed before we can
add that test.  For now, we are maintaining backward compatibility
w.r.t. existing tests.