Add test for TribitsExampleProject2/Package1 raw CMake build (#63)
This simulates the Kokkos native non-TriBITS CMake build system to cover code
that was added in previous commits.

This is is good for a comparison to TriBITS but really this is to test about 5
lines of code added to TribitsInternalPackageWriteConfigFile.cmake to test
this use case (which comes from the Kokkos native non-TriBITS CMake build
system in trilinos/Trilinos#11545).