Add raw non-TriBITS CMake build system for Package1 (#63)
It is incredible how much boiler-plate raw CMake code is needed to create a
simple CMake package with external dependencies.

This is is good for a comparison to TriBITS but really this is to test about 5
lines of code added to TribitsInternalPackageWriteConfigFile.cmake to test
this use case (which comes from the Kokkos native non-TriBITS CMake build
system in trilinos/Trilinos#11545).