Allow TPLs to be listed in LIB_[REQUIRED|OPTIONAL]_PACKAGES (#63)
This required just a minor change in the function tribits_is_pkg_defined() and
removed the argument pkgsOrTpls because it does not matter.  (This will fix
the SEACAS case described in sandialabs/seacas#377.)

I changed the Dependencies.cmake files in the TriBITS projects
TribitsExampleProject, TribitsExampleProject2, and ReducedMockTrilinos to use
the XXX_PACKAGES lists and not the XXX_TPLS list but I left the
Dependencies.cmake in the larger MockTrilinos project alone to make sure we
maintain backward compatibility (even if these are not the best examples for
now).

But I did updaste the code to actually append the XXX_TPLS lists to the
XXX_PACKAGES lists internally so the testing of this should be pretty solid.
(I should likely change all of the example projects to use only the
XXX_PACKAGES list to be a good example.)

See the new entry in the `CHANGLOG.md` file and updated documentation in this
commit for details.

NOTE: I was also able to remove the howto section "how to add a new tribits
external package/tpl dependency" since there is no difference between adding a
dependency on an existing defined upstream internal or external package.

NOTE: This changed the order of internal vs. external packages in the
`<Package>_[TEST|LIB]_DEFINED_DEPENDENCIES` lists which impacted the tests but
does not otherwise impact behavior (because the order if packages in these
lists does not matter).