Print out <Package>_DIR and error out if not set (#63)
This provides useful info about TriBITS-compliant external packages that are
found and it also asserts that <Package>_DIR is non-empty.  It should be
impossible for this to happen but this was observed on a user's machine as
described in:

  https://gitlab.kitware.com/cmake/cmake/-/issues/24667

I was not able to reproduce this issue myself but having this check would have
saved us at least 4 FTE hours and a lot of frustration.