WIP: Add failing system-level test for gitdist with detached head state
I broke the system-level functioning of detailed head info when I refactored
and then squashed the commit:

  gitdist: dist-repo-status: Display tag or SHA1 instead of 'HEAD'

to change how 'showMoreHeadDetails' is passed in to get the CheckinTest.py
tests to pass.  I did not go back and ensure that the 'gitdist' script was
behaving as it should.  I was missing a system-level test to ensure the
correct behavior.

The fix is one-line (which I will add in the next commit) but it took over 100
lines of test code (in this commit) to test this change.  That is terrible.
The tests need to be more maintainable.  I added a todo in the unit tests file
to refactor the testing code to reduce duplication and make the tests more
maintainable.  The way the testing is done now is technical debt and is
starting to look like the CheckinTest.py tests (which is why it is so hard to
modify that code).  I really should have done that refactroing here but it
should be done before any other major work on this code.