Fix system-level bug not displaying the details of HEAD
I broke the system-level functioning of the detached head state details when I
refactored and then squashed the commit:

  gitdist: dist-repo-status: Display tag or SHA1 instead of 'HEAD'

to change how 'showMoreHeadDetails' is passed in to get the CheckinTest.py
tests to pass.  I did not go back and ensure that the 'gitdist' script was
behaving as it should.  I was missing a system-level test to ensure the
correct behavior.

This fix makes the failing test added in the previous commit pass.