Address git describe errors in GitHub Actions jobs
This commit contains changes to try to fix the tests TriBITS_build_docs and
TriBITS_build_sphinx_docs where the `git describe --match="<tag-base-name>*"`
command fails complaining there are not tags in the repo.

It seems that GitHub actions removed tag info with its git clone.  We have to
those tags back because they are used to generate the documentation which
includes the exact version.  A git fetch was done to bring back the tags.
However, that did not seem to work.

This commit also contains error handling for git describe command to produce a
dummy TribitsGitVersion.txt file in case the the git describe command fails.
This seems to have allowed the tests to pass.