Improve documentation for HAVE_<PACKAGE_NAME_UC>_<UPSTREAM_PACKAGE_NAME_UC> (#63)
I noticed this while working on the last commit.

Part of this was changing the name of:

  OPTIONAL_DEP_PACKAGE_NAME

to:

  UPSTREAM_PACKAGE_NAME

in:

  ${PACKAGE_NAME}_ENABLE_${UPSTREAM_PACKAGE_NAME}

Why would anyone try to disable support for a required package so trying to
set ${PACKAGE_NAME}_ENABLE_${UPSTREAM_PACKAGE_NAME}=OFF where
${UPSTREAM_PACKAGE_NAME} is a required upstream dependency of ${PACKAGE_NAME}
makes no sense (but TriBITS sets these vars to 'ON' even for required
dependencies to simplify certain types of logic).