Refactored chunk of tribits_generate_single_repo_version_string for reuse (#597)
Moved git commit info functionality of tribits_generate_single_repo_version_string to
a separate helper function named tribites_generate_commit_info_string. The core functionality and
output is the same. This is in prep for reuse of that functionality inside of the original
tribits_generate_single_repo_version string.