Added initial implementation of cdash random failure tool (#600)
This initial script implementation takes in several cdash arguments
and filters cdash for an initial set of all failing tests for a
certain number of days. With that set of all failing tests, the script
will then get all of that test's testing history. The test's full testing
history is used to build a set of target,topic sha1 associated with failing
testing iterations.

This initial implementation current lacks the check to see if a
passing test's target,topic sha1s exist in the set of failing sha1s,
which denotes an unstable test.

Monolithic commit as this started from a lot of exploratory coding
that eventually built to this starting implementation.