Set -DTribitsExMetaProj_SHOW_GIT_COMMIT_PARENTS=ON for multi-repo test (#597)
This sets the option:

  -DTribitsExMetaProj_SHOW_GIT_COMMIT_PARENTS=ON

in the inner CMake configuration with a multi-repo setup so we can test the
extraction and printing of the parent commit info.  This tests a call of the
function tribits_generate_single_repo_version_string() for extra repos from
the function tribits_generate_repo_version_file_string().  This test also
validates that the correct Git commands are called since this uses real Git on
a Git repos.

The downside of this updated test is that it makes it a little more difficult
to maintain these tests when we have to update these snapshotted Git repos.